knapsack: ship Stage 1 (checker) of the upfront-DAG reimplementation (#200)
Reorganises Knapsack into a subdirectory family. The existing
implementation is preserved as KnapsackLegacy (logic untouched, just
renamed) so we can benchmark the two approaches; a new Knapsack class
follows the BinPacking / MDD pattern with a prepare / define_proof_model
/ install_propagators split.

Stage 1 is the all-fixed checker only: when every item variable is
single-valued, each totals[x] is forced to the computed sum by
JustifyUsingRUP, closing from the natural per-equation OPB constraint
plus the assigned-item literals. No mid-search pruning; Stage 2 will add
the static DAG built once in prepare(), top-level proof scaffolding via
install_initialiser, and a per-call forward+backward sweep.

The knapsack_test data is trimmed to two small 0/1 cases that complete
under a pure checker, and is run via solve_for_tests (no GAC check). A
new knapsack_legacy_test retains the full data set and the GAC check
against KnapsackLegacy. Stage 2 will restore the full coverage for
Knapsack.

Frontends keep posting Knapsack and pick up the new class transparently.
