+++ gcs/constraints/knapsack.hh
@@ -2,5 +2,6 @@
 #define GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_CONSTRAINTS_KNAPSACK_HH
 
 #include <gcs/constraints/knapsack/knapsack.hh>
+#include <gcs/constraints/knapsack/knapsack_legacy.hh>
 
 #endif
