+++ gcs/constraints/knapsack/knapsack_legacy_test.cc
@@ -1,5 +1,5 @@
 #include <gcs/constraints/innards/constraints_test_utils.hh>
-#include <gcs/constraints/knapsack.hh>
+#include <gcs/constraints/knapsack/knapsack_legacy.hh>
 #include <gcs/problem.hh>
 #include <gcs/solve.hh>
 
@@ -8,6 +8,7 @@
 #include <cstdlib>
 #include <functional>
 #include <iostream>
+#include <random>
 #include <set>
 #include <tuple>
 #include <utility>
@@ -26,11 +27,15 @@
 using std::cerr;
 using std::flush;
 using std::make_optional;
+using std::max;
+using std::mt19937;
 using std::nullopt;
 using std::pair;
+using std::random_device;
 using std::set;
 using std::string;
 using std::tuple;
+using std::uniform_int_distribution;
 using std::vector;
 
 #if defined(__cpp_lib_print) && defined(__cpp_lib_format)
@@ -44,16 +49,9 @@ using fmt::println;
 using namespace gcs;
 using namespace gcs::test_innards;
 
-// Stage 1 of the new Knapsack implementation is an all-fixed checker:
-// no mid-search pruning. solve_for_tests (no GAC check) plus a trimmed
-// data set keeps runtimes tractable. KnapsackLegacy retains the full
-// data set and GAC verification — see knapsack_legacy_test.cc. Stage 2
-// will restore wider coverage and switch back to
-// solve_for_tests_checking_gac.
-
-auto run_knapsack_test(bool proofs, pair<int, int> valrange, const vector<vector<int>> & coeffs, const vector<pair<int, int>> & bounds) -> void
+auto run_knapsack_legacy_test(bool proofs, pair<int, int> valrange, const vector<vector<int>> & coeffs, const vector<pair<int, int>> & bounds) -> void
 {
-    print(cerr, "knapsack {} {} {} {}", valrange, coeffs, bounds, proofs ? " with proofs:" : ":");
+    print(cerr, "knapsack_legacy {} {} {} {}", valrange, coeffs, bounds, proofs ? " with proofs:" : ":");
     cerr << flush;
 
     set<tuple<vector<int>, vector<int>>> expected, actual;
@@ -106,24 +104,24 @@ auto run_knapsack_test(bool proofs, pair<int, int> valrange, const vector<vector
         for (const auto & w : coeffs[i])
             coeffs_integers[i].push_back(Integer(w));
 
-    p.post(Knapsack{coeffs_integers, vs, bs});
+    p.post(KnapsackLegacy{coeffs_integers, vs, bs});
 
-    auto proof_name = proofs ? make_optional("knapsack_test") : nullopt;
-    solve_for_tests(p, proof_name, actual, tuple{vs, bs});
+    auto proof_name = proofs ? make_optional("knapsack_legacy_test") : nullopt;
+    solve_for_tests_checking_gac(p, proof_name, expected, actual, tuple{vs, bs});
 
     check_results(proof_name, expected, actual);
 }
 
-// Dup-variable test: Knapsack with a duplicated handle in `vars` (or in
+// Dup-variable test: KnapsackLegacy with a duplicated handle in `vars` (or in
 // `totals`). Two vars positions sharing a handle should be equivalent
 // to one position with summed coefficients; the PB encoding sums by
 // coefficient in normal form. Consistency isn't checked on dup runs;
 // see tmp/duplicate_var_audit.md.
-auto run_dup_knapsack_test(bool proofs, const string & label, pair<int, int> valrange, const vector<pair<int, int>> & unique_var_domains,
+auto run_dup_knapsack_legacy_test(bool proofs, const string & label, pair<int, int> valrange, const vector<pair<int, int>> & unique_var_domains,
     const vector<int> & positions, const vector<vector<int>> & coeffs, const vector<pair<int, int>> & totals_bounds) -> void
 {
-    print(cerr, "knapsack dup {} val={} unique_doms={} positions={} coeffs={} totals={}{}", label, valrange, unique_var_domains, positions, coeffs,
-        totals_bounds, proofs ? " with proofs:" : ":");
+    print(cerr, "knapsack_legacy dup {} val={} unique_doms={} positions={} coeffs={} totals={}{}", label, valrange, unique_var_domains, positions,
+        coeffs, totals_bounds, proofs ? " with proofs:" : ":");
     cerr << flush;
 
     set<tuple<vector<int>, vector<int>>> expected, actual;
@@ -161,9 +159,9 @@ auto run_dup_knapsack_test(bool proofs, const string & label, pair<int, int> val
         for (auto c : row)
             coeffs_i.back().push_back(Integer(c));
     }
-    p.post(Knapsack{coeffs_i, vs, bs});
+    p.post(KnapsackLegacy{coeffs_i, vs, bs});
 
-    auto proof_name = proofs ? make_optional("knapsack_test_dup_" + label) : nullopt;
+    auto proof_name = proofs ? make_optional("knapsack_legacy_test_dup_" + label) : nullopt;
     solve_for_tests(p, proof_name, actual, tuple{unique_vars, bs});
     check_results(proof_name, expected, actual);
     (void)valrange;
@@ -171,12 +169,20 @@ auto run_dup_knapsack_test(bool proofs, const string & label, pair<int, int> val
 
 auto main(int, char *[]) -> int
 {
-    // Cases trimmed to ones that complete quickly under an all-fixed
-    // checker: 0/1 variables with tight total bounds, the (18,19)/(3,8)
-    // corner case, and the degenerate cases. Wider-total and >2-equation
-    // cases are covered by knapsack_legacy_test.
-    vector<tuple<pair<int, int>, vector<vector<int>>, vector<pair<int, int>>>> data = {{{0, 1}, {{2, 3, 4}, {2, 3, 4}}, {{0, 8}, {3, 5}}}, //
-        {{0, 1}, {{7, 4, 4, 7}, {1, 2, 1, 0}}, {{18, 19}, {3, 8}}},                                                                        //
+    vector<tuple<pair<int, int>, vector<vector<int>>, vector<pair<int, int>>>> data = {{{0, 1}, {{2, 3, 4}, {2, 3, 4}}, {{0, 8}, {3, 1000}}}, //
+        {{0, 1}, {{2, 3, 4}, {2, 3, 4}}, {{3, 8}, {3, 1000}}},                                                                                //
+        {{0, 1}, {{2, 3, 4}, {2, 3, 4}}, {{0, 8}, {3, 5}}},                                                                                   //
+        {{0, 1}, {{1, 3, 4}, {2, 0, 8}}, {{0, 8}, {3, 1000}}},                                                                                //
+        {{0, 1}, {{2, 0, 8}, {1, 3, 4}}, {{0, 8}, {3, 1000}}},                                                                                //
+        {{0, 1}, {{2, 0, 8}, {2, 0, 8}}, {{0, 8}, {3, 1000}}},                                                                                //
+        {{0, 1}, {{2, 2, 2, 2, 2}, {2, 2, 2, 2, 2}}, {{0, 5}, {5, 1000}}},                                                                    //
+        {{0, 1}, {{3, 3, 2, 3}, {2, 5, 6, 8}}, {{0, 7}, {4, 1000}}},                                                                          //
+        {{0, 1}, {{8, 2, 4, 3}, {6, 5, 5, 6}}, {{0, 4}, {13, 1000}}},                                                                         //
+        {{0, 1}, {{5, 4, 8, 7}, {2, 5, 1, 5}}, {{0, 12}, {5, 1000}}},                                                                         //
+        {{0, 1}, {{8, 7, 4, 8}, {4, 3, 4, 4}}, {{0, 18}, {10, 1000}}},                                                                        //
+        {{0, 1}, {{7, 4, 4, 7}, {1, 2, 1, 0}}, {{18, 19}, {3, 8}}},                                                                           //
+        {{2, 4}, {{4, 1, 2, 3}, {4, 6, 3, 8}, {5, 3, 1, 6}}, {{0, 64}, {0, 48}, {0, 41}}},                                                    //
+        {{1, 4}, {{1, 8, 3, 1}}, {{3, 42}}},                                                                                                  //
         // Degenerate cases (issue #254).
         // Empty item list (zero columns): every weighted sum is 0, so the
         // totals must admit 0.
@@ -190,18 +196,34 @@ auto main(int, char *[]) -> int
         {{1, 1}, {{2, 3}}, {{5, 5}}},  // all taken 1: sum 5 == total (tautology)
         {{1, 1}, {{2, 3}}, {{4, 4}}}}; // all taken 1: sum 5 != total 4 (contradiction)
 
+    random_device rand_dev;
+    mt19937 rand(rand_dev());
+    for (int x = 0; x < 10; ++x) {
+        uniform_int_distribution n_coeffs_dist(1, 4), size_dist(1, 4), item_dist(0, 8), bound_dist(0, 40), delta_dist(0, 30);
+        auto size = size_dist(rand);
+        auto n_coeffs = n_coeffs_dist(rand);
+
+        vector<pair<int, int>> boundses;
+        for (int k = 0; k < n_coeffs; ++k) {
+            auto wb = bound_dist(rand) + 20;
+            boundses.emplace_back(max(0, wb - (25 + delta_dist(rand))), wb);
+        }
+
+        generate_random_data(rand, data, random_bounds(0, 2, 1, 3), vector(n_coeffs, vector(size, item_dist)), boundses);
+    }
+
     for (bool proofs : {false, true}) {
         if (proofs && ! can_run_veripb())
             continue;
         for (auto & [valrange, coeffs, bounds] : data)
-            run_knapsack_test(proofs, valrange, coeffs, bounds);
+            run_knapsack_legacy_test(proofs, valrange, coeffs, bounds);
 
         // {x, x} — single var doubled; coefficients sum (1+2)*x = 3x and
         // (2+3)*x = 5x; totals must equal those.
-        run_dup_knapsack_test(proofs, "xx", {0, 3}, {{0, 3}}, {0, 0}, {{1, 2}, {2, 3}}, {{0, 9}, {0, 15}});
+        run_dup_knapsack_legacy_test(proofs, "xx", {0, 3}, {{0, 3}}, {0, 0}, {{1, 2}, {2, 3}}, {{0, 9}, {0, 15}});
 
         // {x, y, x} — x's coefficient is summed at positions 0 and 2.
-        run_dup_knapsack_test(proofs, "xyx", {0, 2}, {{0, 2}, {0, 2}}, {0, 1, 0}, {{1, 2, 1}, {3, 1, 2}}, {{0, 8}, {0, 12}});
+        run_dup_knapsack_legacy_test(proofs, "xyx", {0, 2}, {{0, 2}, {0, 2}}, {0, 1, 0}, {{1, 2, 1}, {3, 1, 2}}, {{0, 8}, {0, 12}});
     }
 
     return EXIT_SUCCESS;
